Source code cleanup and optimizations

+++ src/Hangfire.Mongo/MongoWriteOnlyTransaction.cs
@@ -96,6 +96,11 @@ namespace Hangfire.Mongo
 
         public override void IncrementCounter(string key)
         {
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x => x.StateData.InsertOne(new CounterDto
             {
                 Id = ObjectId.GenerateNewId(),
@@ -106,6 +111,11 @@ namespace Hangfire.Mongo
 
         public override void IncrementCounter(string key, TimeSpan expireIn)
         {
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x => x.StateData.InsertOne(new CounterDto
             {
                 Id = ObjectId.GenerateNewId(),
@@ -117,6 +127,11 @@ namespace Hangfire.Mongo
 
         public override void DecrementCounter(string key)
         {
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x => x.StateData.InsertOne(new CounterDto
             {
                 Id = ObjectId.GenerateNewId(),
@@ -127,6 +142,11 @@ namespace Hangfire.Mongo
 
         public override void DecrementCounter(string key, TimeSpan expireIn)
         {
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x => x.StateData.InsertOne(new CounterDto
             {
                 Id = ObjectId.GenerateNewId(),
@@ -143,6 +163,11 @@ namespace Hangfire.Mongo
 
         public override void AddToSet(string key, string value, double score)
         {
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             var builder = Builders<SetDto>.Update;
             var set = builder.Set(_ => _.Score, score);
             var setTypesOnInsert = builder.SetOnInsert("_t", new[] { nameof(KeyValueDto), nameof(ExpiringKeyValueDto), nameof(SetDto) });
@@ -162,6 +187,11 @@ namespace Hangfire.Mongo
 
         public override void RemoveFromSet(string key, string value)
         {
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x => x.StateData
                 .OfType<SetDto>()
                 .DeleteMany(
@@ -171,6 +201,11 @@ namespace Hangfire.Mongo
 
         public override void InsertToList(string key, string value)
         {
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x => x.StateData.InsertOne(new ListDto
             {
                 Id = ObjectId.GenerateNewId(),
@@ -181,6 +216,11 @@ namespace Hangfire.Mongo
 
         public override void RemoveFromList(string key, string value)
         {
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x => x.StateData
                 .OfType<ListDto>()
                 .DeleteMany(
@@ -190,21 +230,27 @@ namespace Hangfire.Mongo
 
         public override void TrimList(string key, int keepStartingFrom, int keepEndingAt)
         {
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x =>
             {
-                int start = keepStartingFrom + 1;
-                int end = keepEndingAt + 1;
-
-                ObjectId[] items = ((IEnumerable<ListDto>)x.StateData.OfType<ListDto>()
-                        .Find(new BsonDocument())
-                        .Project(Builders<ListDto>.Projection.Include(_ => _.Key))
-                        .Project(_ => _)
-                        .ToList())
+                var items = x.StateData
+                    .OfType<ListDto>()
+                    .Find(new BsonDocument())
+                    .Project<KeyValueDto>(Builders<ListDto>.Projection.Include(_ => _.Id))
+                    .ToEnumerable()
+                    .Where((data, i) => (i < keepStartingFrom) || (i > keepEndingAt))
+                    .Select(data => data.Id)
                     .Reverse()
-                    .Select((data, i) => new { Index = i + 1, Data = data.Id })
-                    .Where(_ => ((_.Index >= start) && (_.Index <= end)) == false)
-                    .Select(_ => _.Data)
                     .ToArray();
+
+                if (!items.Any())
+                {
+                    return;
+                }
 
                 x.StateData
                     .OfType<ListDto>()
@@ -216,10 +262,14 @@ namespace Hangfire.Mongo
         public override void SetRangeInHash(string key, IEnumerable<KeyValuePair<string, string>> keyValuePairs)
         {
             if (key == null)
+            {
                 throw new ArgumentNullException(nameof(key));
+            }
 
             if (keyValuePairs == null)
+            {
                 throw new ArgumentNullException(nameof(keyValuePairs));
+            }
 
             var builder = Builders<HashDto>.Update;
             var setTypesOnInsert = builder.SetOnInsert("_t", new[] { nameof(KeyValueDto), nameof(ExpiringKeyValueDto), nameof(HashDto) });
@@ -250,7 +300,9 @@ namespace Hangfire.Mongo
         public override void RemoveHash(string key)
         {
             if (key == null)
+            {
                 throw new ArgumentNullException(nameof(key));
+            }
 
             QueueCommand(x => x.StateData.OfType<HashDto>().DeleteMany(Builders<HashDto>.Filter.Eq(_ => _.Key, key)));
         }
@@ -269,15 +321,16 @@ namespace Hangfire.Mongo
         }
 
 
-
-        //New methods to support Hangfire pro feature - batches.
-
-
+        // New methods to support Hangfire pro feature - batches.
 
 
         public override void ExpireSet(string key, TimeSpan expireIn)
         {
-            if (key == null) throw new ArgumentNullException(nameof(key));
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x => x
                 .StateData
                 .OfType<SetDto>()
@@ -287,7 +340,11 @@ namespace Hangfire.Mongo
 
         public override void ExpireList(string key, TimeSpan expireIn)
         {
-            if (key == null) throw new ArgumentNullException(nameof(key));
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x => x.StateData
                 .OfType<ListDto>()
                 .UpdateMany(Builders<ListDto>.Filter.Eq(_ => _.Key, key),
@@ -296,7 +353,11 @@ namespace Hangfire.Mongo
 
         public override void ExpireHash(string key, TimeSpan expireIn)
         {
-            if (key == null) throw new ArgumentNullException(nameof(key));
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x => x.StateData
                 .OfType<HashDto>()
                 .UpdateMany(Builders<HashDto>.Filter.Eq(_ => _.Key, key),
@@ -305,7 +366,11 @@ namespace Hangfire.Mongo
 
         public override void PersistSet(string key)
         {
-            if (key == null) throw new ArgumentNullException(nameof(key));
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x => x.StateData
                 .OfType<SetDto>()
                 .UpdateMany(Builders<SetDto>.Filter.Eq(_ => _.Key, key),
@@ -314,7 +379,11 @@ namespace Hangfire.Mongo
 
         public override void PersistList(string key)
         {
-            if (key == null) throw new ArgumentNullException(nameof(key));
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x => x.StateData
                 .OfType<ListDto>()
                 .UpdateMany(Builders<ListDto>.Filter.Eq(_ => _.Key, key),
@@ -323,7 +392,11 @@ namespace Hangfire.Mongo
 
         public override void PersistHash(string key)
         {
-            if (key == null) throw new ArgumentNullException(nameof(key));
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x => x.StateData
                 .OfType<HashDto>()
                 .UpdateMany(Builders<HashDto>.Filter.Eq(_ => _.Key, key),
@@ -332,8 +405,14 @@ namespace Hangfire.Mongo
 
         public override void AddRangeToSet(string key, IList<string> items)
         {
-            if (key == null) throw new ArgumentNullException(nameof(key));
-            if (items == null) throw new ArgumentNullException(nameof(items));
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+            if (items == null)
+            {
+                throw new ArgumentNullException(nameof(items));
+            }
             var builder = Builders<SetDto>.Update;
 
 
@@ -363,7 +442,10 @@ namespace Hangfire.Mongo
 
         public override void RemoveSet(string key)
         {
-            if (key == null) throw new ArgumentNullException(nameof(key));
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
             QueueCommand(x => x.StateData
                 .OfType<SetDto>()
                 .DeleteMany(Builders<SetDto>.Filter.Eq(_ => _.Key, key)));
